django/camac/dossier_import/config/kt_ag/requirements.txt
hdbcli==2.23.24
python-dotenv==1.0.1
